PresenterTests added and presenter functionality is implemented including Favourite Mode support

+++ Movies/Core/Models/Favorite+Methods.swift
@@ -11,16 +11,16 @@ import CoreData
 
 extension Favorite {
     
-    static func getInstance(uid: String, insertNewIfNeeded: Bool = false) -> Favorite? {
+    static func getInstance(film_uid: String, insertNewIfNeeded: Bool = false) -> Favorite? {
         do {
             let fetchRequest = NSFetchRequest<NSFetchRequestResult>(entityName: String(describing: Favorite.self))
-            fetchRequest.predicate = NSPredicate(format: "film_uid = %@", uid)
+            fetchRequest.predicate = NSPredicate(format: "film_uid = %@", film_uid)
             if let fetchResults = try DataController.getContext().fetch(fetchRequest) as? [Favorite] {
                 if fetchResults.count > 0 {
                     return fetchResults[0]
                 } else if insertNewIfNeeded {
                     let favorite = NSEntityDescription.insertNewObject(forEntityName: String(describing: Favorite.self), into: DataController.getContext()) as? Favorite
-                    favorite?.film_uid = uid
+                    favorite?.film_uid = film_uid
                     return favorite
                 } else {
                     return nil
@@ -35,6 +35,24 @@ extension Favorite {
     
     static func deleteAll() {
         for favorite in Favorite.getAll() ?? [Favorite]() {
+            DataController.getContext().delete(favorite)
+        }
+    }
+    
+    static func toggleFavorite(film_uid: String) {
+        if getInstance(film_uid: film_uid) != nil {
+            Favorite.deleteInstance(film_uid: film_uid)
+        } else {
+            _ = Favorite.getInstance(film_uid: film_uid, insertNewIfNeeded: true)
+        }
+    }
+    
+    static func isFavorite(film_uid: String) -> Bool {
+        return getInstance(film_uid: film_uid) != nil
+    }
+    
+    static func deleteInstance(film_uid: String) {
+        if let favorite = getInstance(film_uid: film_uid) {
             DataController.getContext().delete(favorite)
         }
     }
